Added notvalid routes and small clean up
