+++ src/index.ts
@@ -1,12 +1,11 @@
-import { MongoClient, ServerApiVersion } from 'mongodb';
-import { Client as ElasticClient } from '@elastic/elasticsearch';
-
-
 import * as dotenv from 'dotenv';
+import * as log4js from 'log4js';
 import express from 'express';
 import { getMongoClient } from './databases/mongo';
 import { getElasticSearchClient } from './databases/elasticSearch';
 import { configureLogger } from './utils/logger';
+import { configureNotValidRoute, debugRequest } from './utils/requests';
+import { APPLICATION_CONFIG } from './utils/applicationConfig';
 
 dotenv.config();
 
@@ -20,21 +19,27 @@ const elasticSearchApiKey = process.env.ELASTICSEARCH_API_KEY || '';
 
 async function main() {
     configureLogger();
+
+    const logger = log4js.getLogger("Main");
+
     try {
+        const elasticSearchClient = await getElasticSearchClient(elasticSearchCloudId, { id: elasticSearchId, api_key: elasticSearchApiKey });
+
         const mongoClient = await getMongoClient(mongoUri);
-        const elasticSearchClient = await getElasticSearchClient(elasticSearchCloudId, {id: elasticSearchId, api_key: elasticSearchApiKey});        
 
-        await elasticSearchClient.ping();
+        configureNotValidRoute(app);
 
+        if(APPLICATION_CONFIG.DEBUG_REQUEST === true){ 
+            debugRequest(app);
+        }
+
+        app.listen(port, () => {
+            logger.log(`Server is running on port ${port}`);
+        });
 
     } catch (error) {
-        
+        logger.error("error during connection", error);
     }
-
-    
-
-
-
 }
 
 main().catch((error) => {
